app/requirements_web.txt
# Web Application Requirements for Best Wines Sweden

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# HTTP client (existing)
httpx==0.25.2

# HTML parsing (existing)
selectolax==0.3.27

# Async support
anyio==3.7.1

# Data processing
python-multipart==0.0.6

# Template engine
jinja2==3.1.2

# Existing dependencies (from current requirements.txt)
retrying==1.3.4
unidecode==1.3.8

# Vivino scraper dependencies
selectolax==0.3.27

# AI services for intelligent wine matching
google-generativeai==0.3.2
openai==1.35.0

# Optional: For development
pytest==7.4.3
pytest-asyncio==0.21.1

# Pydantic for data validation
pydantic==2.5.0

# For serving static files
aiofiles==23.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4